more styling changes and some file changes as well
