--- Frontend/src/components/Navbar/NavbarData.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import * as FaIcons from 'react-icons/fa';
-
-const NavbarData = [
- {path:'#About', text: 'About', icon: <FaIcons.FaHome />},
- {path:'#Experience', text: 'Experience', icon: <FaIcons.FaHome />},
- {path:'#Projects', text: 'Projects', icon: <FaIcons.FaAddressCard />},
- {path:'#Contact', text: 'Contact', icon: <FaIcons.FaAddressBook />},
-]
-
-export default NavbarData
