+++ Frontend/src/components/Navbar/styles.ts
@@ -83,14 +83,6 @@ export const NavItemLink = styled(LinkS)`
  padding: 0 1rem;
  height: 100%;
  cursor: pointer;
- 
- &.active {
-  border-bottom: 3px solid #01bf71;
- }
- 
- &:hover {
-  color: #01bf71;
- }
 `;
 
 export const NavResumeButton = styled.nav`
