+++ Frontend/src/components/index.ts
@@ -3,7 +3,6 @@ export { default as Layout} from './Layout/Layout';
 export { default as LeftSidebar} from './LeftSidebar/LeftSidebar';
 export { default as RightSidebar} from './RightSidebar/RightSidebar';
 export { default as Navbar } from './Navbar/Navbar';
-export { default as NavbarData } from './Navbar/NavbarData'; 
 export { default as NavbarLogo } from './Navbar/Logo/Logo';
 export { default as NavbarMenuToggle } from './Navbar/NavbarMenuToggle/NavbarMenuToggle';
 export { default as Footer } from './Footer/Footer';
